Refactor LogMessageArchiveTest*
